Add Id annotation and refactor package structure to cleanly separate models;
